Rock, Paper, Scissors game
I've made rock, paper, scissors games plenty of times before, and I've
been trying to follow the instructions on TOP so far. I had a pretty
weird time trying to make the logic for this game work following the
stated instructions so it took me more time than usual.

I'm happy with how it turned out, but I probably wouldn't do it like
this in the future.